UI: Add attack section to RansomwareReport
